As read on http://www.cs.ucsb.edu/~rich/class/cs290I-grid/notes/Sockets/ :
There are two subtleties that you should notice. First, the call to
siginterrupt() is important. It tells the operating system that system
calls should not be restarted after a signal is received by the process.
On some systems (including many versions of Linux), the default action is
to go back and restart the system call (thereby negating the signal's
ability to interrupt and cause a time out).


git-svn-id: https://svn.opendtect.org/od/trunk@5160 a985855d-d384-4b45-ae33-31a410998c3c