Enabled automatic computation (as well as computation without the need to plot the curve) of rms error as soon as user draws a line.
Summary:
a) Fixed the crash that occurs when entered equation results in a huge rms error.
b) Made the default range correspond to the one that is used when the crossplot is autoscaled with the current clip ratio.
c) Made the drawn line to be removed from the crossplot scene when the user clicks cancel button.
d) Created a function that updates latest range and clipping option in the scaling tab whenever it is notified of a data change by crossplotter.
e) Made the input field of the rms error 'read only.'

Test Plan: All changes correspond to crossplot properties dialog box.

Reviewers: raman, satyaki

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com/D15

git-svn-id: https://svn.opendtect.org/od/trunk@28459 a985855d-d384-4b45-ae33-31a410998c3c