Fix for event-filter dead-locks.
Summary: The event-filter receives all events from qt. The problem is that it triggers an notifier, and the actions by the notifier may trigger more events, which leads to a dead-lock. Thix fix removes the mutex-lock as soon as the protected code is over (i.e. the use of qobj_).

Test Plan: Review logic.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com/D36

git-svn-id: https://svn.opendtect.org/od/trunk@28463 a985855d-d384-4b45-ae33-31a410998c3c