Added NaN check in mIsUdf(float_complex).
Summary:
Normally, when od is compiles with __debug__ enabled, the macro mIsUdf(float) and mIsUdf(double) calls the dbgIsUdf(float) or dbgIsUdf(double) which checks for NaN numbers.

This fix should expand that check to float_complex as well.

Test Plan: Not tested (only compiles & links). A breakpoint should be set in dbgIsUdf( float_complex ) and ensure that it is run in every mIsUdf() with float_complex.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D41

git-svn-id: https://svn.opendtect.org/od/trunk@28472 a985855d-d384-4b45-ae33-31a410998c3c