getColor() function of fault and polygonbody display returns darker color than selected by the user.
Summary:
At re-entrance, the properties window does not show your originally selected
color (the one used for the sticks/polygons), but a darker one (the one
used for the surfaces). This bugfix is a revision of -r19373 (fault & body
color icons did not refresh until extra click in tree). Earlier, getColor()
returned activestickmarker_->getMaterial()->getColor(), but that is wrong
as well as long as activestickmarker_->getMaterial()->setColor(.) in the
function updateSingleColor() is preceded by getMaterial()->setColor(.),
which immediately triggers a call to getColor(). Safest solution is to have
getColor() return nontexturecol_.

Test Plan:
Tested for faults and polygon bodies, changing the color in their display
properties multiple times, while keeping an eye on their color in the 
3D scene and at their tree item. Visual review should suffice.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D72

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28588 a985855d-d384-4b45-ae33-31a410998c3c