Display material properties lost when restoring a MultiTextureSurveyObject
Summary:
When restoring a session with MultiTextureSurveyObjects like an inline,
a fault, etc., its display material properties like ambient and diffuse
reflection and transparency are lost. Solved if the fillPar() and usePar()
functions of the MultiTextureSurveyObject call the appropriate base class.

Test Plan:
Tested for restored session with inline and fault. Please comment if you
see any reason against calling VisualObjectImpl instead of VisualObject.
My concern is that the old situation existed since 2006.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D90

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28615 a985855d-d384-4b45-ae33-31a410998c3c