Fault texture stays black after restoring a session.
Summary:
Fault texture stays black after restoring a session, because the attribute
number is derived from the one selected in the tree. Nothing is selected
in the tree when restoring a session.

Test Plan:
I tested by loading a fault, adding an attribute, saving the session and
restoring it. Please comment if you had a reason not to use the "attrib"
parameter, but the visServer()->getSelAttribNr() instead, which happens to
return -1 if the function is called via a session restore. You may ignore
the last conditional statement I added. It belongs to new functionality.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D91

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28622 a985855d-d384-4b45-ae33-31a410998c3c