Changed port in .arcconfig
Summary: Same as D80, but in od4.4. As our policy allows commits in od_osg, osgGeo and od_qt (for another week), I will bluntly commit there. For od4.4, phabricatior will automatically issue a post-commit review of the change if it does not go through the review. You can expect the same for dgb4.4 in a few minutes.

Test Plan: Visual inpection.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D83

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28636 a985855d-d384-4b45-ae33-31a410998c3c