CmdRecStopper involved in crashes, because its stack did pop without push
Summary:
Turns out that when removing tree or list items, the CmdRecStopper may 
receive a zero object pointer. It was refused when pushed on the stack,
but still popped later on. This was already incorrect, but is now causing
a crash in /od, since the pop() function no longer checks for empty stack.

Test Plan:
The workflow that caused an occasional crash was run with valgrind. The
many reported memory violations with a CmdRecStopper destructor in the
stack dump are all gone after this bugfix.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D147

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28721 a985855d-d384-4b45-ae33-31a410998c3c