Quick fix for data root issue on Unix
Summary: If DTECT_DATA is set to '/', OD keeps asking for data root with each restart. This has to be fixed eventually in the Installer where DTECT_DATA should be taken as user input. But for now we need a quick fix that is safe.

Test Plan: Tested on Linux32.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D137

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28740 a985855d-d384-4b45-ae33-31a410998c3c