If manager is deliberately disabled in canvas right-click menu, it could still be popped up by double-click on canvas.
Summary:
It seems one is not supposed to get access to the color table manager
when selecting for example a color blending attribute in the tree. It is
consistent to ignore double-click if manager is not in right-click menu.

Test Plan:
Selected a color blending attribute and found that double click in canvas
no longer pops up manager, while it still does for normal attributes.

Reviewers: satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D158

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28745 a985855d-d384-4b45-ae33-31a410998c3c