Moved test-programs to src directory structure.
Summary: I do not want the test-programs to be more visible to the developers, but I don't want it to pollute the src/<Module>/ directory. My compromize is to create a src/<Module>/tests directory where the programs are located. The knowledge of the test-programs are then added to the normal src/<Module>/CMakeLists.txt file.

Test Plan: Tested on Linux, and all test programs are built in the right place, and works.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D162

git-svn-id: https://svn.opendtect.org/od/trunk@28757 a985855d-d384-4b45-ae33-31a410998c3c