Split part of the FSS-to-fault functionality for (1) fault import and (2) fault stick transfer by the faulttoolmanager.
Summary:
Bugfix: When transferring faultsticks to a fault via the fault toolbar, 
warnings telling that some sticks cannot be incorporated pop up regularly.
New intelligence for importing faults (August 2012) screws up, because
faulttoolmanager expects the sticks to stay invariant and not merged, etc.
Restored old functionality for this case, because the user knows and sees
what he/she is doing when using the fault toolbar and does not need its
sticks to be re-edited.

Test Plan:
Checked that fault import and fault stick transfer each call their own 
functionality, and that the regular warnings in the latter case are gone.
Also checked that there are no other parts of OD calling this functionality!

Reviewers: yuancheng

Reviewed By: yuancheng

Differential Revision: https://phabricator.dgbes.com:8443/D176

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28775 a985855d-d384-4b45-ae33-31a410998c3c