Added the possibility to set the send application for a crash.
Summary:
In batch programs, we do not always have the possibility to pop up a ui, so we
need to send reports without a ui. Hence, it should be possible to steer
which application to use. I also added the --binary flag for the reporter, 
which will be picked up by my next revision of the od_ReportIssue.

Test Plan:
Tested by running an executable with ui things and make sure it first sets
it to non-ui, and then ui.

Reviewers: aneesh

Reviewed By: aneesh

Maniphest Tasks: T23

Differential Revision: https://phabricator.dgbes.com:8443/D181

git-svn-id: https://svn.opendtect.org/od/trunk@28779 a985855d-d384-4b45-ae33-31a410998c3c