Updated issue reporter with new commandline parser.
Summary:
The issue reporter has change to incorporate the new command-line parser. Changes also include a new --binary flag that must be used for sending non-text files. The --binary flag is implemented in D181.

The change also sends all files to the standard crashreport.php script, which 
must be modified to get this fully working. That change is a part of T23

Test Plan: Tested od_ReportIssue and checked that the parsing of all options was fine.

Reviewers: aneesh

Reviewed By: aneesh

Maniphest Tasks: T23

Differential Revision: https://phabricator.dgbes.com:8443/D183

git-svn-id: https://svn.opendtect.org/od/trunk@28780 a985855d-d384-4b45-ae33-31a410998c3c