Putting a level in the display as much as possible
Summary:
Now separated LayerSequence::indexOf(level) which does a strict lookup of the level. What we want in display is to see the level where it is across hiatuses, including at the top or bottom  of the sequence if that applies.
Because it's just adding (and using) 2 new non-virtual functions this doesn't break ABI.

Test Plan: Tried this in od first, then transplanted and tested again. Made a model with 3 layers, added levels above, inside and below the generated layers.

Reviewers: satyaki

Reviewed By: satyaki

CC: helene

Differential Revision: https://phabricator.dgbes.com:8443/D179

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28781 a985855d-d384-4b45-ae33-31a410998c3c