Removed Plugins::setArgs().
Summary:
All OpendTect programs are mandated to set the argc and argv through
SetProgramArgs as the first thing in the main function. Hence, there is no need
to set it on the Plugins class, as the class can retrieve them from the system.

Test Plan: od_main starts, as well as od_process_attrib. The other execuables have been inspected so that their corresponding main function calls SetProgramArgs.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D200

git-svn-id: https://svn.opendtect.org/od/trunk@28802 a985855d-d384-4b45-ae33-31a410998c3c