Added removeSequence() to LayerModel
Summary:
To enable editing layer models, we alo need a removeSequence().
While at it transferred some fixes to read/write of layer models from od. This functionality is not used in od4.4, but may be later.

Test Plan:
Tested removeSequence in od.
read/write are now for the first time working in od, so can only be better than the totally untested stuff that was in od4.4.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D216

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28851 a985855d-d384-4b45-ae33-31a410998c3c