Intersection marker fix in HorizonCube Editor.
Summary:
Problem:Intersection markers on horizons  in HorizonCube Editor are shown for pair of inline and crossline. After moving inline/crossline to new positions intersection markers for previous pair are shown even if no seeds have been picked on them.
Fix:Intersection markers on horizons  in HorizonCube Editor  will now be shown only on horizons and inline or crossline on which seeds have been picked.

Test Plan:
Tested with F3 Demo
Load a inline and crossline in the scene. Add HC display on inline/crossline.
Launch HC control center and go for tools->edit.
In HC editor add a event to inline and pick seeds on it. Then moves the inline to a new position and again moves it to new position. Intersection marker on crossline will be seen only for the inline position on which seeds have been picked. In this case only one marker is shown.
Also tested by picking seeds on crossline and moving crossline or picking seeds on both inline/crossline and moving them in different combinations.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D256

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28932 a985855d-d384-4b45-ae33-31a410998c3c