Changes in initialization of LogSampler
Summary:
Removed all mentions to SI().zStep() since the zstep is a mandatory argument of the function
The checks previously done on line 1039, 1043 will not work if zrgintime_ = true
Put the computation of dah in a macro
Removed use of mLocalEps since 0.1 is too close to a usual step value of 0.15. Now using an epsilon that is 1% of zstep
Replaced lines 1064-1103 by the macro mGetZ. By using it I define zrgstep that is in the extraction domain.
Rounded off zstart, zstop to a multiple of the step
Moved the computation of winsz from doLog() into lines 1075-1087 (new version) since that is common to all logs. winsz is then stored is a new component of data_, at the end for compatibility purposes. Then it can be read in doLog().

Test Plan:
Tested from the well tie module that uses this log sampler (example with zrgintime_ = true and extrintime_ = true)
Tested from the SynthRock-profile module that uses this log sampler (example with zrgintime_ = false and extrintime_ = false)
Tested with create-log cube in time survey with extraction boundaries in depth
Tested with create-log cube in depth survey with extraction boundaries in depth

Reviewers: salil

Reviewed By: salil

Differential Revision: https://phabricator.dgbes.com:8443/D258

git-svn-id: https://svn.opendtect.org/od/trunk@28935 a985855d-d384-4b45-ae33-31a410998c3c