Fixed deadlock when attaching own notifiers.
Summary: When using the new mAttachCB (rather than notifier.notify() ), there was a bug that made the object add itself as listener. This in turn gave a dead lock when destructing the object.

Test Plan: Tested on mac. The resolution is that we do not allow callbacks to attach to themselfs, and they are hence never removed from themselves.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D270

git-svn-id: https://svn.opendtect.org/od/trunk@28950 a985855d-d384-4b45-ae33-31a410998c3c