Bugfix #0001790, No rotation axis in scene after restore session
Summary: After restoring a session, there's no rotation axis in the scene, though the button in the toolbar is active

Test Plan: Restored an empty scene, and the axis is there. The axis display status is not stored in the session itself, so the axis is now always visible after a restore session.

Reviewers: ranojay

Reviewed By: ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D249

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@28977 a985855d-d384-4b45-ae33-31a410998c3c