Changes to ODGraphicsDynamicImageItem
Summary: Fix for mantis bug with ID 1799. Changing colortable sequence does not update the bitmap display in 2D Viewer if the change is done after zooming in. This is because currently only bitmaps of surrounding low resolution part of uiDynamicImageItem are being updated. This can be noticed by zooming out a little after zooming in and changing the colortable sequence in the 2D Viewer.

Test Plan: Tested by changing the color sequence after zooming in on 2D Viewer in Variable Density mode.

Reviewers: nanne, raman, kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D221

git-svn-id: https://svn.opendtect.org/od/trunk@28995 a985855d-d384-4b45-ae33-31a410998c3c