Added possibility to set allow CommandLineParser::getVal() to return true if key is not there.
Summary:
There was one functionality missing, and that is if you want to allow something like:

program [--file <mandatory>]

you could not do that. It is perfectly OK if --file is not in place, but if it is, we must be able to parse what comes after.

Test Plan: Added cases to test program, and they function fine.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D308

git-svn-id: https://svn.opendtect.org/od/trunk@29023 a985855d-d384-4b45-ae33-31a410998c3c