Fixed a crash
Summary: OpendTect crashes if the callbacker is dynamic-casted to uiODViewer2D after the 2D Viewer is deleted/removed. This is because viewer2DWinClosedCB is currently called twice if the closed 2D Viewer was in both Wiggle and VD mode.

Test Plan: Tested by closing the 2D Viewer in both Wiggle and VD mode.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D346

git-svn-id: https://svn.opendtect.org/od/trunk@29085 a985855d-d384-4b45-ae33-31a410998c3c