Made refcounts completely threadsafe.
Summary: There was a glitch in the refcount implementation that made it possible, if timing was unlucky. That a one thread could ref an object while it was being deleted. Thanks to our Atomics, we could close that gap.

Test Plan: Compiles & Links. New code is walked through in the debugger. The code is used extensively, so the fact that OD at all starts with the change is a good test.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D324

git-svn-id: https://svn.opendtect.org/od/trunk@29090 a985855d-d384-4b45-ae33-31a410998c3c