Bug fix in restore session: Depth scene scaling problem.
Summary:
Restoring a session with depth scene resulted in a scaling problem on the depth scene. It was occuring as zscale was not set properly.
Port of D178 to include this bug fix in next OD4.4(i) patch realease.

Test Plan:
Tested with time survey with depth in meter. Restored session with depth scene, wheeler scene, and normal scene.
Tested with time survey with depth in feet. Restored session with depth scene, wheeler scene, and normal scene.
Tested with depth survey. Restored session with time scene, wheeler scene, and normal scene.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D342

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@29112 a985855d-d384-4b45-ae33-31a410998c3c