Fix for launching multiple instances of Multi-machine processing (od_SeisMMProc.exe) mantis ID 1351
Summary: As per the implementation in 4.0, the socket provider keeps on trying with the next port number with an increment of one till the max number(3000) is reached. The same logic is applied with the new implementation, where if it fails for the first port, it increases the value with 1 and tries with the new port value. Also added error message display if the socket fails to listen, which was not there in 4.0

Test Plan: Tested with launching multiple instances of Multi-machine processing with different jobs for each, and all completed successfully.

Reviewers: nanne

Reviewed By: nanne

CC: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D348

git-svn-id: https://svn.opendtect.org/od/trunk@29145 a985855d-d384-4b45-ae33-31a410998c3c