Experimental usage of std::atomic.
Summary:
std::atomic is what we should have had 5 years ago, but it comes with c++11. Once it is in place, 90% of our atomic code can go. Now, most of us don't have it, and, as a matter of fact, I don't either.

The std::atomic also comes with some adaptions of the Atomic::setIfEqual concept, which is the reason I divided it int strongSetIfEqual and weakSetIfEqual.

Test Plan:
The c++11 code is not even compiled. The detection in plfdefs.h is taken from
sourceforge. The test program for the atomic class is however extensive, and covers all functions. So the day our compiler does c++11, we will know for sure if it works (or even compiles).

The test program is also extended with a stress-test, where two threads are started that constantly change the atomic, while the main thread tries to do an setIfEqual. It should reasonably both fail and succeed during 10M attempts, depending on the synchronization with the other threads.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D372

git-svn-id: https://svn.opendtect.org/od/trunk@29166 a985855d-d384-4b45-ae33-31a410998c3c