Bug fixes in restore session:Depth scene scaling problem. Port of D342 to OD
Summary:
Restoring a session with depth scene resulted in a scaling problem on the depth scene. It was occuring as zscale was not set properly.
Also for zaxistransformed scene z-scale was not saved in .sess file.

Test Plan:
Tested with time survey with depth in meter. Restored session with depth scene, wheeler scene, and normal scene.
Tested with time survey with depth in feet. Restored session with depth scene, wheeler scene, and normal scene.
Tested with depth survey. Restored session with time scene, wheeler scene, and normal scene.
Also tested with session saved while running OD4.4 and then restoring them with OD as session saved with OD4.4 does not save zscale in .sess file.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D387

git-svn-id: https://svn.opendtect.org/od/trunk@29187 a985855d-d384-4b45-ae33-31a410998c3c