Added dummy usage of variables set on the test command line.
Summary:
When cmake configures od from the release/test scripts, it sets -DISTAG and -DBRANCH on the command line. As these are not used in some projects, it causes a warning in cdash.

I have added dummy usage of these variables to avoid the warning.

Test Plan: CMake works. No functional change.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D415

git-svn-id: https://svn.opendtect.org/od/trunk@29242 a985855d-d384-4b45-ae33-31a410998c3c