Added new class of test programs that should only be run nightly.
Summary: Test programs that take long time to execute should not be run in the continuous integration, as they take long time. Hence the OD_NIGHTLY_TEST_PROGS variable was added to facilitate that.

Test Plan: Not tested as the first program does not exist yet, and the the infrastructure is not in place yet. CMake works without problems.

Reviewers: aneesh

Reviewed By: aneesh

Maniphest Tasks: T27

Differential Revision: https://phabricator.dgbes.com:8443/D449

git-svn-id: https://svn.opendtect.org/od/trunk@29265 a985855d-d384-4b45-ae33-31a410998c3c