Template for angle_computer.cc
Summary:
This is a skeleton of a test program to test the angle computer. My idea is that you (initially) hardcode a multiID to read from the F3 data, as well as a hard-coded TraceID.

You would then compute the angles, and do a few checks at a few depths/offset that the angles are correct.

Later, I want to do a few things:

1. I wish to add put the MultiID and the TraceID in a parameter file.
2. I wish to use the test-survey that Nageswara and Arnaud are making (T15).

Please also note that when D124 is approved, the executable will be named test_angle_computer.

Test Plan: This does not compile. It is my wish that you commandeer this revision (so it becomes yours) and implement it. I'll be happy to review it for you when you reach that stage.

Reviewers: kristofer

Reviewed By: kristofer

Maniphest Tasks: T8

Differential Revision: https://phabricator.dgbes.com:8443/D130

git-svn-id: https://svn.opendtect.org/od/trunk@29267 a985855d-d384-4b45-ae33-31a410998c3c