Introduced the case when avilable and desired ranges are different
Summary: Made changes for the case when desired and available zrange are different. We basically try to take the intersection of two layers.

Test Plan: Tested with F3 Demo and F3_Test_Survey. In case of F3_Test_Survey availablez is larger than desiredz. Tested the case when desiredz is larger than availablez by manually changing the survey parameter in .survey file.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D483

git-svn-id: https://svn.opendtect.org/od/trunk@29339 a985855d-d384-4b45-ae33-31a410998c3c