Fixed wrong order in steps taken during well tie
Summary:
moved the sonic integration to the first step of the process: the d2ModelMgr must be created before WellTie::Data. As a result Data can no longer be an argument, but only the well tie setup is actually needed.
Also the extraction range needed to be updated after each change of the time-depth model and that was missing.

Test Plan: tested using various datasets, especially Eagle Ford

Reviewers: aneesh

Reviewed By: aneesh

Differential Revision: https://phabricator.dgbes.com:8443/D507

git-svn-id: https://svn.opendtect.org/od/trunk@29420 a985855d-d384-4b45-ae33-31a410998c3c