Defined equal operators for strings.
Summary:
When comparing a const char* to a BufferString, the wrong operator was used. Hence, the

  bool operator==(const char*, const BufferString& )
  bool operator!=(const char*, const BufferString& )

These are not implemented, as we don't want to risk promoting to BufferString in the wrong places.
 
Then, a couple ambigious casts required us to implement.

  bool FixedString::operator==(const BufferString&) const
  bool FixedString::operator!=(const BufferString&) const

Test Plan: Compiles and Links on mac. The dgb repository compiles and links. OpendTect starts. I have not been able to run through FixedString::operator== and !=, but they should be so straight forward that I don't deem it necessary.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D522

git-svn-id: https://svn.opendtect.org/od/trunk@29438 a985855d-d384-4b45-ae33-31a410998c3c