Added a function.
Summary: To avoid memory leakage. Corresponding datapacks are not released if the display of wva/vd is set to off before closing the 2D Viewer. This is because datapacks are currently released using packID(wva) which currently returns zero if the display is off.

Test Plan: DPM(4).packs_.size() is zero after closing all 2D Viewers and removing displayed inlines/crosslines in the 3D scene.

Reviewers: satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D577

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@29546 a985855d-d384-4b45-ae33-31a410998c3c