Port of D619 to OD4.6:Bug Fix for mantis bug id 1850
Summary:
Active HorizonCube does not remember the polygon name in the HC creator.
If a HorizonCube is computed within a polygon, it remembers the P:ranges only in the HC creator. It does not remember the polygon name. The field remains empty.
Steps To Reproduce:Active HorizonCube (select any HC that is computed within a polygon).
1-Create HorizoNCube
2- Area subselection - press Select
3- In the pop-up window you will see no polygon.

Test Plan: Followed steps to reproduce the bug. Now polygon will be shown but only for HC's that are created after this fix.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D640

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@29600 a985855d-d384-4b45-ae33-31a410998c3c