2D Viewer closes when properties dialog is opened.
Summary: Displayed a well in the scene. Displayed the well in a 2D Viewer. Opened the properties dialog. 2D Viewer closed automatically before the properties dialog is popped up. This should not happen as properties are to be changed while the 2D Viewer is open.

Test Plan: Now the 2D Viewer doesn't close when the properties dialog is opened.

Reviewers: aneesh, raman

Reviewed By: aneesh

Differential Revision: https://phabricator.dgbes.com:8443/D696

git-svn-id: https://svn.opendtect.org/od/trunk@29672 a985855d-d384-4b45-ae33-31a410998c3c