Introducing Z skip for display
Summary:
When displaying layer models and synthetics, tehre are often long stretches of uninteresting depth ranges that should not be displayed. Added a function to set this value on various base classes.

After this, we need uiStratSynthDisp to make use of the new 'dispskipz_'.

Test Plan: I tested this but while uiStratSynthDisp doesn't respond to the new variable, nothing really changes - as uiStratSynthDisp sets the zoom. This of course is good enough for now, as it means no deterioration.

Reviewers: satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D720

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@29728 a985855d-d384-4b45-ae33-31a410998c3c