Bugfix for erronious lines (in axes and other places)
Summary: Sometimes, the tick-lines on the axis in the flatviewer have been in the wrong position. The reason for that is that the position (origin) of the line has been set previously, and when the line is re-used as a tick-line, the old origin was not reset.

Test Plan: Tested on normal axis-displays, and works.

Reviewers: mahant

Reviewed By: mahant

Differential Revision: https://phabricator.dgbes.com:8443/D782

git-svn-id: https://svn.opendtect.org/od/trunk@29831 a985855d-d384-4b45-ae33-31a410998c3c