Porting memory leak fix (D791) into od4.6 without interface changes
Summary:
The uiODFaultToobox had an auxcolorinput_ field. That field had no parent,
and hence the deletion of it went wrong in that it did not delete the qt
parts, leading to a memory leak.

Test Plan:
I can create faults, and they get the selected color. The color button is
correclty updated. Switched from used-defined color mode to random color
mode and back. All looks fine.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D817

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@29842 a985855d-d384-4b45-ae33-31a410998c3c