Fixed memory leak in the fault-toolbox.
Summary:
The uiODFaultToobox had an auxcolorinput_ field. That field had no parent, and hence the deletion of it went wrong in that it did not delete the qt parts, leading to a memory leak.

I replaced the auxcolorinput_ with a call to selectColor.

Test Plan: It seems to work fine. I can create faults, and they get the selected color. The color button is correclty updated.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D791

git-svn-id: https://svn.opendtect.org/od/trunk@29843 a985855d-d384-4b45-ae33-31a410998c3c