Shifted release of DataPacks.
Summary: Datapacks are not released in cases where 2D Viewer is deleted, but not closed. Hence shifted their release to the destructor which is called in both cases.

Test Plan: Tested by removing the corresponding treeitem in OpendTect main window while the 2D Viewer is open and after it is closed. In both cases DPM(4).packs_.size() was zero at the end.

Reviewers: satyaki, raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D865

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@29897 a985855d-d384-4b45-ae33-31a410998c3c