Fix for bug 1889, z-scale not correctly set for linear t2d transforms.
Summary:
The zScale implementation was missing on Linear T2D transforms. Solved
by making base implementation smarter. Also did minor renaming of zScale to toZScale to make it clear that this is the scale for the 'to' domain.

Test Plan: Tested for time and depth surveys with depth and time scenes.

Reviewers: salil

Reviewed By: salil

Differential Revision: https://phabricator.dgbes.com:8443/D900

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@29958 a985855d-d384-4b45-ae33-31a410998c3c