Port of D900 to od.
Summary: See D900 for reasons. Change is brought a bit further here as we are not in a stable release: The implementation of zScale in T2D objects are removed, as the default implementation is good enough.

Test Plan: Tested on depth and time surveys, with both time and depth scenes.

Reviewers: salil

Reviewed By: salil

Differential Revision: https://phabricator.dgbes.com:8443/D902

git-svn-id: https://svn.opendtect.org/od/trunk@29961 a985855d-d384-4b45-ae33-31a410998c3c