Bug fix : Random line display in 2D Viewer
Summary: If you display a random line in 2D viewer when the random line is partly outside the data range, the display is blank. Basically the size of PosData() and arr2d_ should match. This fix does exactly that.

Test Plan: Tested by displaying random lines with partially filled data in 2D viewer.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D946

git-svn-id: https://svn.opendtect.org/od/trunk@30011 a985855d-d384-4b45-ae33-31a410998c3c