Shifted release of DataPacks.
Summary: DataPacks are not released in cases where 2D Viewer is deleted, but not closed. Hence shifted their release to the destructor which is called in both cases. Also made changes so that appropriate display parameters are read before handleChange is called.

Test Plan: Tested by removing the corresponding treeitem in OpendTect main window while the 2D Viewer is open and after it is closed. In both cases DPM(4).packs_.size() was zero at the end. Also now, there is no double handleChange ( or atleast doesn't appear as if there is one ) when an attribute that is displayed in the 3D scene is displayed in the 2D Viewer as the display parameters are read before handleChange is called.

Reviewers: satyaki, raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D971

git-svn-id: https://svn.opendtect.org/od/trunk@30032 a985855d-d384-4b45-ae33-31a410998c3c