Fix for T49 - crash in WCP
Summary:
It turns out that the callbacks to the well-data is never removed. Hence the crash as the notifier tries to call a dead object.

As the multi-well version has no own destructor, the cb cannot be removed there, and the single-well destructor won't do as it is too late for in the multi-well case.

I added two onClose functions, to attach to when the window is closed, but they are not always triggered. Hence onClose will be pubic and called before deletion from the dgb-plugin.

Test Plan: Comiles and works. Crash is gone.

Reviewers: nanne

Reviewed By: nanne

Maniphest Tasks: T49

Differential Revision: https://phabricator.dgbes.com:8443/D937

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@30062 a985855d-d384-4b45-ae33-31a410998c3c