Setting the default size of string representation to precision.
Summary:
Arnaud reports in Mantis bug #1749 that to few decimals are written by OD, and that it should be 7 valid digits for float, and 15 for doubles.
Only worry is that GUI strings may get very long, but the trimming seems to work, so no problems have been observed. If it becomes a problem, we would have to have separate toString functions when writing to disk.

Test Plan: Compiles and works.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D910

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30095 a985855d-d384-4b45-ae33-31a410998c3c