Disable base color controls in set material property dialog
Summary:
In current od versions, if the body is visualized by color tabel, the base color controls still are 
active in material property dialog. The user can play it but it doesn't influence the visualization. 
On purpose, we shouldn't change the base color as Arnaud said. Then I think we should disable the controls to avoid 
the misunstanding by clients.

Test Plan: Compiles and Tested in body display.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1140

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30277 a985855d-d384-4b45-ae33-31a410998c3c