od: Fix for typed input not being used
Summary: As a fallout of D1124, one could not type in an output cube name and use it. The idea behind the fix is to limit the setAttrNm call which was eventually clearing the input field.

Test Plan: Tested in both 2D and 3D with both typed and selected output cubes.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D1231

git-svn-id: https://svn.opendtect.org/od/trunk@30354 a985855d-d384-4b45-ae33-31a410998c3c