od: avoiding false hostname_
Summary: The stremprovider looks for the character ':' and if found, everything before that is treated as the hostname. I have a command (from GMT) where an argument can have the character ':' and in that case you cannot extract hostname_ like this. This fix assumes (rightly so I guess) that a hostname cannot have spaces.

Test Plan: It parses my example command correctly.

Reviewers: bert

Reviewed By: bert

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1244

git-svn-id: https://svn.opendtect.org/od/trunk@30373 a985855d-d384-4b45-ae33-31a410998c3c