Fixed problem with AtomicPointer type on windows.
Summary: The atomic pointer type was set to od_uint64, which made the atomic use the generic, non-specialized construct. This lead to crashes.

Test Plan: Tested in od and od4.6. The bug is not present in od4.6 yet, as the AtomicPointer is not used in the SpinLock. That is however pending.

Reviewers: salil

Reviewed By: salil

Differential Revision: https://phabricator.dgbes.com:8443/D1257

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30392 a985855d-d384-4b45-ae33-31a410998c3c