Added threads.h include on windows.
Summary:
When the locker was introduced, the include to 'threads.h' was removed from atomic. On windows however, there is only native support for int atomics on win32, and int and 64-bit int on win64. In OD, we rely on both 32 and 64 bit ints to be
present on all platforms, and hence we will have to use mutex for 64-bit ints on win32. Hence, the include is needed until we use std::atomic.

Test Plan: Compile warnings goes away.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D1286

git-svn-id: https://svn.opendtect.org/od/trunk@30428 a985855d-d384-4b45-ae33-31a410998c3c