Making the new Threads::Lock/Locker available in 4.6
Summary:
The new Lock(er) will be in od code from now on, so it's better for porting efforts to have the object available.
To get it there the recursive locks need to be available so had to take the latest versions of the thread objects and test programs.

Test Plan: The test programs test_threads test_threadwork still succeed, and OpendTect starts and does things.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1287

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30429 a985855d-d384-4b45-ae33-31a410998c3c